added ClearDatabaseService.kt to clean the code
