+++ app/src/main/java/pl/llp/aircasting/database/ClearDatabaseService.kt
@@ -1,0 +1,19 @@
+package pl.llp.aircasting.database
+
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.async
+import kotlinx.coroutines.runBlocking
+
+class ClearDatabaseService {
+    fun clearDatabase() {
+        // to make sure downloading sessions stopped before we start deleting them
+        Thread.sleep(1000)
+        runBlocking {
+            val query = GlobalScope.async(Dispatchers.IO) {
+                DatabaseProvider.get().clearAllTables()
+            }
+            query.await()
+        }
+    }
+}
