+++ app/src/main/java/pl/llp/aircasting/screens/settings/clear_sd_card/my_account/MyAccountController.kt
@@ -10,12 +10,14 @@ import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.async
 import kotlinx.coroutines.runBlocking
 import org.greenrobot.eventbus.EventBus
+import pl.llp.aircasting.database.ClearDatabaseService
 
 class MyAccountController(
     private val mContext: Context,
     private val mViewMvc: MyAccountViewMvc,
     private val mSettings: Settings
 ) : MyAccountViewMvc.Listener{
+    private val clearDatabaseService = ClearDatabaseService()
 
     fun onStart(){
         mViewMvc.registerListener(this)
@@ -30,19 +32,9 @@ class MyAccountController(
         EventBus.getDefault().post(LogoutEvent())
 
         mSettings.logout()
-        clearDatabase()
+        clearDatabaseService.clearDatabase()
 
         LoginActivity.startAfterSignOut(mContext)
     }
 
-    private fun clearDatabase() {
-        // to make sure downloading sessions stopped before we start deleting them
-        Thread.sleep(1000)
-        runBlocking {
-            val query = GlobalScope.async(Dispatchers.IO) {
-                DatabaseProvider.get().clearAllTables()
-            }
-            query.await()
-        }
-    }
 }
